changed 3rd parameter of gaussaw() to size_t in gird.h and after_sptrans.c, added cast in Importbinary.c

+++ src/Importbinary.c
@@ -92,7 +92,7 @@ int y_is_gauss(double *gridyvals, int ysize)
       double *yvals, *yw;
       yvals = (double*) malloc(ysize*sizeof(double));
       yw    = (double*) malloc(ysize*sizeof(double));
-      gaussaw(yvals, yw, ysize);
+      gaussaw(yvals, yw, (size_t)ysize);
       free(yw);
       for ( i = 0; i < (int) ysize; i++ )
 	yvals[i] = asin(yvals[i])/M_PI*180.0;
@@ -138,7 +138,7 @@ int define_grid(dsets_t *pfi)
 
   if ( pfi->yrflg ) rev_vals(yvals, ny);
 
-  if ( pfi->linear[1] == 0 ) lgauss = y_is_gauss(yvals, ny);
+  if ( pfi->linear[1] == 0 ) lgauss = y_is_gauss(yvals, (size_t)ny);
 
   if ( lgauss ) gridtype = GRID_GAUSSIAN;
   else          gridtype = GRID_LONLAT;
